Move some boot-related info to pre-kernel area
