+++ Boot/CMakeLists.txt
@@ -1,6 +1,15 @@
+LIST(APPEND BOOT_HEADERS DeviceTree/DeviceTree.hpp)
+LIST(APPEND BOOT_SOURCES DeviceTree/DeviceTree.cpp)
+
+LIST(APPEND BOOT_HEADERS PSCI/PSCI.hpp)
+LIST(APPEND BOOT_SOURCES PSCI/PSCI.cpp)
+
+LIST(APPEND BOOT_HEADERS BootDriver.hpp)
+LIST(APPEND BOOT_SOURCES BootDriver.cpp)
+
 LIST(APPEND BOOT_SOURCES Boot.cpp)
 
-ADD_LIBRARY(Boot STATIC ${BOOT_SOURCES})
+ADD_LIBRARY(Boot STATIC ${BOOT_HEADERS} ${BOOT_SOURCES})
 TARGET_LINK_LIBRARIES(Boot ArchLib)
 
 SET_TARGET_PROPERTIES(Boot PROPERTIES LINK_DEPENDS ${LINKER_SCRIPT})
