+++ Devices/kern_drivers.hpp
@@ -2,14 +2,10 @@
 #include <kern_runtime.hpp>
 #include <kern_datatypes.hpp>
 
-#include <PhyProtocol/DeviceTree/DeviceTree.hpp>
 
 #ifndef _KERN_DRIVERS_HPP_
 #define _KERN_DRIVERS_HPP_
 
-void InitDriver(const CHAR *DriverName, UINT64 Address);
-void DriverHandleDTB(const CHAR *DriverName, DeviceTreeBlob *CurState);
-void FiniDriver(const CHAR *DriverName, UINT64 Address);
 
 void PerCoreInit();
