+++ System/PhyProtocol/CMakeLists.txt
@@ -1,11 +1,5 @@
-LIST(APPEND PHY_PROTOCOL_HEADERS DeviceTree/DeviceTree.hpp)
-LIST(APPEND PHY_PROTOCOL_SOURCES DeviceTree/DeviceTree.cpp)
-
 LIST(APPEND PHY_PROTOCOL_HEADERS PCI/PCIe.hpp)
 LIST(APPEND PHY_PROTOCOL_SOURCES PCI/PCIe.cpp)
-
-LIST(APPEND PHY_PROTOCOL_HEADERS PSCI/PSCI.hpp)
-LIST(APPEND PHY_PROTOCOL_SOURCES PSCI/PSCI.cpp)
 
 ADD_LIBRARY(PhyProtocol STATIC
 	${PHY_PROTOCOL_HEADERS}
